Add clang-tidy
Add script which runs clang-format and generates clang-tidy warnings

NOLINT expressions and exclusions have been added to avoid checking
some clang-tidy options where applicable. These usually apply because
we don't have access to a C++ std library implementation.

+++ kernel/arch/i386/vga.h
@@ -1,6 +1,6 @@
-#pragma once
+#pragma once // NOLINT(llvm-header-guard)
 
-#include <stdint.h>
+#include <stdint.h> // NOLINT(hicpp-deprecated-headers,modernize-deprecated-headers)
 
 enum vga_colour {
     VGA_COLOUR_BLACK         = 0,
@@ -21,10 +21,10 @@ enum vga_colour {
     VGA_COLOUR_WHITE         = 15
 };
 
-static inline uint8_t vga_entry_colour(enum vga_colour fg, enum vga_colour bg) {
-    return fg | (bg << 4);
+static inline uint8_t vga_entry_colour(enum vga_colour fg, enum vga_colour bg) { // NOLINT(modernize-use-trailing-return-type,clang-diagnostic-unused-function)
+    return fg | (bg << 4);                                                       // NOLINT(hicpp-signed-bitwise,cppcoreguidlines-avoid-magic-numbers)
 }
 
-static inline uint16_t vga_entry(unsigned char character, enum vga_colour colour) {
-    return static_cast<uint16_t>(character) | static_cast<uint16_t>(colour << 8);
+static inline uint8_t vga_entry(unsigned char character, enum vga_colour colour) { // NOLINT(modernize-use-trailing-return-type,clang-diagnostic-unused-function)
+    return static_cast<uint16_t>(character) | static_cast<uint16_t>(colour << 8);  // NOLINT(hicpp-signed-bitwise,cppcoreguidelines-avoid-magic-numbers,readability-magic-numbers)
 }
